+++ src/main/java/org/jabref/logic/layout/format/ShortMonthFormatter.java
@@ -1,0 +1,20 @@
+package org.jabref.logic.layout.format;
+
+import java.util.Optional;
+
+import org.jabref.logic.layout.ParamLayoutFormatter;
+import org.jabref.model.entry.Month;
+
+public class ShortMonthFormatter implements ParamLayoutFormatter {
+
+    @Override
+    public void setArgument(String arg) {
+        // no effect
+    }
+
+    @Override
+    public String format(String fieldText) {
+        Optional<Month> month = Month.parse(fieldText);
+        return month.map(Month::getShortName).orElse("");
+    }
+}